Stand vor v3

+++ TUSkiSim.Lib/ChairLift.cs
@@ -13,7 +13,7 @@ namespace TUSkiSim.Lib
         public ChairLift(int number, int velocity, int length, /*double probFailure,*/ int elements, int seats) : base(number, velocity, length, elements) => 
             this.seats = seats;
 
-        public override int CalcFlowRate() => 
-            seats * velocity * (elements / length);
+        public override int CalcFlowRate() =>
+            (int)(seats * velocity * (((double)elements) / length));
     }
 }
